+++ examples/ex_display_brightness.py
@@ -14,7 +14,7 @@ disp_peaks = np.logspace(np.log10(1), np.log10(1000), 5)
 
 # Display parameters
 contrast = 1000   # Display contrast 1000:1
-gamma = 2.2       # Standard gamma-encoding
+EOTF = "2.2"       # Standard gamma-encoding
 E_ambient = 100   # Ambient light = 100 lux
 k_refl = 0.005    # Reflectivity of the display
 
@@ -22,7 +22,7 @@ metric = pycvvdp.cvvdp(display_name='standard_4k', heatmap='threshold')
 
 Q_JOD = []
 for dd, Y_peak in enumerate(disp_peaks):
-    disp_photo = pycvvdp.vvdp_display_photo_eotf(Y_peak=Y_peak, contrast=contrast, EOTF='sRGB', gamma=gamma, E_ambient=E_ambient, k_refl=k_refl)
+    disp_photo = pycvvdp.vvdp_display_photo_eotf(Y_peak=Y_peak, contrast=contrast, EOTF=EOTF, E_ambient=E_ambient, k_refl=k_refl)
     metric.set_display_model(display_photometry=disp_photo)
 
     q, stats = metric.predict(I_test_noise, I_ref, dim_order="HWC")