+++ ChangeLog.md
@@ -3,5 +3,6 @@
 * Fixed freezing on Windows when reading long videos (due to bug in python's /dev/null implementation)
 * Added plain psnr_rgb metric, which operates on display-encoded values (or PU21-encoded if needed)
 * Updated PU-PSNR-* to use 100 nit as the peak + to computed PSNR for all pixels in the video (rather than mean PSNR over all frames).
+* Reorganized display_models.json - now color space is a part of the display model json spec, EOTF is a part of color space json spec.
 
 # v0.1.0 Initial beta release (27 March 2023)